FoldsThreads.jl

FoldsThreads.FoldsThreads — Module

FoldsThreads: Extra threaded executors for JuliaFolds/*.jl

FoldsThreads.jl provides extra thread-based executors usable with various JuliaFolds/*.jl packages such as Transducers.jl and FLoops.jl.

                                  Executors
                           ,----------------------.
     Algorithms            |    FoldsThreads.jl    |         Data structures
,------------------.       |-----------------------|       ,-----------------.
|  FLoops,         |       |  ThreadedEx*          |       |  Array,         |
|  Folds,          |       |  WorkStealingEx,      |       |  Tables,        |
|  Transducers,    |  ---  |  DepthFirstEx,        |  ---  |  FGenerators,   |
|  OnlineStats,    |       |  TaskPoolEx,          |       |  Dict,          |
|  DataTools, ...  '       |  NondeterministicEx,  |       |  Set, ...       |
`------------------'       |  ...                  |       `-----------------'
                           `-----------------------'

(* ThreadedEx is the default executor provided by Transducers.jl)

WorkStealingEx implements work stealing (continuation stealing). Useful for load-balancing.

DepthFirstEx implements depth-first scheduling. Useful for findfirst-type computations.

TaskPoolEx: Task pool executor. Useful for fine execution control (e.g., back pressure and "background" threads).

NondeterministicEx: An executor for parallelizing computations with non-parallelizable iterators.

FoldsThreads.DepthFirstEx — Type

DepthFirstEx(; [simd,] [basesize])

Depth-first scheduling for parallel execution. Useful for findfirst-type of computation.

Examples

julia> using FoldsThreads
       using Folds

julia> Folds.sum(i -> gcd(i, 42), 1:1000_000, DepthFirstEx())
4642844

Extended help

DepthFirstEx schedules chunks of size roughly equal to basesize in the ordered that each chunk appears in the input collection. The basecase computation does not wait for all the tasks to be scheduled. This approach performs better than a more naive approach where the all tasks are scheduled before the reduction starts. DepthFirstEx is useful for reductions that can terminate early (e.g., findfirst, @floop with break).

Keyword Arguments

basesize: The size of base case.

simd: false, true, :ivdep, or Val of one of them. If true/:ivdep, the inner-most loop of each base case is annotated by @simd/@simd ivdep. Use a plain loop if false (default).

More examples

julia> using FoldsThreads
       using FLoops

julia> @floop DepthFirstEx() for x in 1:1000_000
           y = gcd(x, 42)
           @reduce(acc += y)
       end
       acc
4642844

FoldsThreads.NondeterministicEx — Type

NondeterministicEx(; [simd,] [basesize,] [ntasks,])

Batched parallel reduction for non-parallelizable collections (e.g., Channel, Iterators.Stateful).

This is a simple wrapper of NondeterministicThreading transducer. Use NondeterministicThreading directly for explicit control on what transducers are parallelized.

Examples

julia> using FoldsThreads
       using Folds

julia> partially_parallelizable(seq) = (gcd(y, 42) for x in seq for y in 1:10000x);

julia> Folds.sum(partially_parallelizable(Iterators.Stateful(1:100)), NondeterministicEx())
234462500

In the above example, we can run gcd(y, 42) (mapping), for y in 1:10000x (flattening), and + for sum (reduction) in parallel even though the iteration of Iterators.Stateful(1:100) is not parallelizable. Note that, as indicated in the example, the computation per each iteration of the non-parallelizable collection should be very CPU-intensive in order for NondeterministicEx to show any kind of performance benefits.

Same computation using FLoops.jl:

julia> using FoldsThreads
       using FLoops

julia> @floop NondeterministicEx() for x in Iterators.Stateful(1:100)
           for y in 1:10000x
               z = gcd(y, 42)
               @reduce(acc += z)
           end
       end
       acc
234462500

FoldsThreads.TaskPoolEx — Type

TaskPoolEx(; [simd,] [basesize,] [ntasks,] [background,])

Executor using pooled tasks for reduction. Useful for reductions with I/O and managing back pressure. With background = true, it can also be used to isolate throughput-oriented tasks and use the primary thread for latency-oriented tasks.

Examples

julia> using FoldsThreads
       using Folds

julia> Folds.sum(i -> gcd(i, 42), 1:1000_000, TaskPoolEx())
4642844

Extended help

Worker tasks are pooled (for each executor) so that the number of Julia Tasks used for a reduction can be much smaller than input_length ÷ basesize. This strategy is used mainly for limiting resource (e.g., memory) required by the reduction than for load-balancing. WorkStealingEx performs better for load-balancing of compute-intensive reductions.

NOTE: This executor is inspired by ThreadPools.jl. The hack for assigning a task to a dedicated thread is stolen from ThreadPools.jl.

Warning

It is highly discouraged to use this executor in Julia packages; especially those that are used as libraries rather than end-user applications. This is because the whole purpose of this executor is to prevent Julia runtime from doing the right thing for managing tasks. Ideally, the library user should be able to pass an executor as an argument so that your library function can be used with any executors including TaskPoolEx.

Keyword Arguments

background = false: Do not run tasks on threadid() == 1.

ntasks: The number of tasks to be used.

basesize: The size of base case.

simd: false, true, :ivdep, or Val of one of them. If true/:ivdep, the inner-most loop of each base case is annotated by @simd/@simd ivdep. Use a plain loop if false (default).

More examples

julia> using FoldsThreads
       using FLoops

julia> @floop TaskPoolEx() for x in 1:1000_000
           y = gcd(x, 42)
           @reduce(acc += y)
       end
       acc
4642844

FoldsThreads.WorkStealingEx — Type

WorkStealingEx(; [simd,] [basesize])

Work-stealing scheduling for parallel (but not concurrent) execution. Useful for load-balancing.

Examples

julia> using FoldsThreads
       using Folds

julia> Folds.sum(i -> gcd(i, 42), 1:1000_000, WorkStealingEx())
4642844

Extended help

WorkStealingEx implements work stealing scheduler for Transducers.jl and other JuliaFolds/*.jl packages. Worker tasks are pooled (for each executor) so that the number of Julia Tasks used for a reduction can be much smaller than input_length ÷ basesize. This has a positive impact for reduction that requires load-balancing since this does not incur the overhead of spawning tasks. However, as the worker tasks are occupied by a base case until the base case is fully reduced, the user functions (reducing functions and transducers) cannot use concurrency primitives such as channels and semaphores to communicate within them. See below for discussion on usable concurrency patterns.

NOTE: WorkStealingEx is more experimental than the default multi-thread executor ThreadedEx. Importantly, WorkStealingEx still does not perform well than ThreadedEx for parallel computation that does not require load-balancing.

Keyword Arguments

basesize: The size of base case.

simd: false, true, :ivdep, or Val of one of them. If true/:ivdep, the inner-most loop of each base case is annotated by @simd/@simd ivdep. Use a plain loop if false (default).

More examples

julia> using FoldsThreads
       using FLoops

julia> @floop WorkStealingEx() for x in 1:1000_000
           y = gcd(x, 42)
           @reduce(acc += y)
       end
       acc
4642844

Possible concurrency primitive usages

Each channel is used solely for consuming or producing items (not both):

User functions that only consumes items from channels that are produced by Tasks outside the reduction.

User functions that only produces items to channels that have enough buffer size or are consumed by Tasks outside the reduction.

Locks that are acquired and released within an iteration.